+++ biotrack/track.py
@@ -1,178 +1,172 @@
 # biotrack, Apache-2.0 license
 # Filename: biotrack/track.py
-# Description:  Basic track object to contain and update tracks
+# Description:  Basic track object to contain and update_pt_box tracks
 from collections import defaultdict
-from biotrack.logger import info, debug
+from biotrack.logger import info, err
+from biotrack.trace import Trace
 import numpy as np
 
 
-def max_score_p(model_predictions, model_scores):
-    """Find the top prediction"""
-    max_score = 0.0
-
-    for row in zip(model_predictions, model_scores):
-        prediction, score = row
-        score = float(score)
-        if score > max_score:
-            max_score = score
-            best_pred = prediction
-
-    return best_pred, max_score
-
-
 class Track:
-    def __init__(self, track_id: int, label: str, pt: np.array, emb: np.array, frame: int, x_scale: float, y_scale: float, box: np.array = None, score: float = 0., **kwargs):
+    NUM_KP = 4 # Number of key points to track
+
+    def __init__(self, track_id: int, x_scale: float, y_scale: float, **kwargs):
         max_empty_frames = kwargs.get("max_empty_frames", 30)
         max_frames = kwargs.get("max_frames", 300)
-        info(f"Creating tracker {track_id} at {frame}:{pt},{score}. Max empty frame {max_empty_frames} Max frames {max_frames}")
+        info(f"Creating tracker {track_id}. Max empty frame {max_empty_frames} Max frames {max_frames}")
+        self.traces = [Trace(track_id, i) for i in range(Track.NUM_KP)]
         self.max_empty_frames = max_empty_frames
         self.max_frames = max_frames
+        self.frames = defaultdict(int)
         self.id = track_id
-        self.pt = {frame: pt}
-        self.label = {frame: label}
-        self.score = {frame: score}
-        self.box = {frame: box}
-        self.emb = emb
-        self.best_label = label
-        self.best_score = score
-        self.start_frame = frame
-        self.last_updated_frame = frame
         self.closed = False
         self.x_scale = x_scale
         self.y_scale = y_scale
+        self.emb = []
 
     @property
     def track_id(self):
         return self.id
 
     @property
+    def num_frames(self):
+        # Get the longest trace
+        longest_trace = self.traces[0]
+        for trace in self.traces:
+            if trace.num_frames > longest_trace.num_frames:
+                longest_trace = trace
+        return longest_trace.num_frames
+
     def embedding(self):
         return self.emb
+
+    def get_traces(self):
+        return self.traces
 
     def is_closed(self) -> bool:
         return self.closed
 
-    def close(self, frame_num: int) -> bool:
-        # If the acceleration is very low, then the object being tracked is likely to be stationary
-        # allow the track to remain open for a longer period of time if it is
-        pts = np.array([pt for pt in self.pt.values()])
-        frames = np.array([int(frame) for frame in self.pt.keys()])
-        acceleration_mag = 100.
-        if len(pts) > 2:
-            delta_pos = np.diff(pts, axis=0)
-            delta_time = np.diff(frames)
-            velocity = delta_pos / delta_time[:, np.newaxis]
-            acceleration = np.diff(velocity, axis=0) / delta_time[1:, np.newaxis]
-            acceleration_mag = np.linalg.norm(acceleration[-1]) # Get the last acceleration
-            info(f"Tracker {self.id} acceleration {np.round(acceleration_mag*1000)}")
-        is_closed = ((frame_num - self.last_updated_frame + 1) >= self.max_empty_frames or
-                     len(self.pt) >= self.max_frames and acceleration_mag < 3.)
-        info(f"Tracker {self.id} is_closed {is_closed} frame_num {frame_num} last_updated_frame {self.last_updated_frame} "
+    def dump(self):
+        info(f"================Tracker {self.id} dump================")
+        for trace in self.traces:
+            trace.dump()
+        info(f"=====================================================")
+
+    def close_check(self, frame_num: int) -> bool:
+        # Find the longest trace and get the average accuracy
+        longest_trace = self.traces[0]
+        for trace in self.traces:
+            if trace.num_frames > longest_trace.num_frames:
+                longest_trace = trace
+
+        avg_acc = longest_trace.compute_acc()
+        last_updated_frame = longest_trace.last_update_frame
+
+        is_closed = ((frame_num - last_updated_frame + 1) >= self.max_empty_frames or
+                     longest_trace.num_frames >= self.max_frames and avg_acc < 10. and avg_acc != -1)
+        info(f"{self.id} is_closed {is_closed} frame_num {frame_num} last_updated_frame {last_updated_frame} "
              f"max_empty_frame {self.max_empty_frames} max_frames {self.max_frames}")
         self.closed = is_closed
 
-    @property
-    def last_update_frame(self):
-        return self.last_updated_frame
-
-    def rescale(self, pt: np.array, box: np.array) -> (np.array, np.array):
+    def rescale_pt(self, pt: np.array) -> np.array:
         pt_rescale = pt.copy()
         pt_rescale[0] = pt[0] * self.x_scale
         pt_rescale[1] = pt[1] * self.y_scale
-        if box is not None:
-            box_rescale = box.copy()
-            box_rescale[0] = box[0] * self.x_scale
-            box_rescale[1] = box[1] * self.y_scale
-            box_rescale[2] = box[2] * self.x_scale
-            box_rescale[3] = box[3] * self.y_scale
-        else:
-            box_rescale = box
-        return pt_rescale, box_rescale
+        return pt_rescale
 
-    @property
-    def num_frames(self):
-        return self.last_updated_frame - self.start_frame + 1
+    def rescale_box(self, box: np.array) -> np.array:
+        box_rescale = box.copy()
+        box_rescale[0] = box[0] * self.x_scale
+        box_rescale[1] = box[1] * self.y_scale
+        box_rescale[2] = box[2] * self.x_scale
+        box_rescale[3] = box[3] * self.y_scale
+        return box_rescale
 
     def get_best(self, rescale=True) -> (int, np.array, str, np.array, float):
-        # Get the best box which is a few frames behind the last_updated_frame
-        # This is pretty arbitrary, but sometimes the last box is too blurry or not visible
-        num_frames = len(self.pt.keys())
-        if num_frames > 3:
-            frame_num = list(self.pt.keys())[-3]
-            box = self.box[frame_num]
-            pt = self.pt[frame_num]
-        else: # Handle the case where there is only one frame tracked
-            frame_num = self.last_updated_frame
-            box = self.box[frame_num]
-            pt = self.pt[frame_num]
-        if rescale:
-            pt, box = self.rescale(pt, box)
-        return frame_num, pt, self.best_label, box, self.best_score
+        # Get the longest trace
+        longest_trace = self.traces[0]
+        for trace in self.traces:
+            if trace.num_frames > longest_trace.num_frames:
+                longest_trace = trace
 
-    def get(self, frame_num: int, rescale=True) -> (np.array, str, np.array, float):
-        if frame_num not in self.pt.keys():
-            return None, None, None, 0.
-        pt = self.pt[frame_num]
-        # If there is a box in the frame, return it
-        if self.box[frame_num] is not None:
-            box = self.box[frame_num]
-        else:
-            box = []
+        frames = longest_trace.get_frames()
+        if len(frames) == 0:
+            return None, None, None, None, 0.
+
+        # Find a frame somewhere near the middle of the trace that has a box
+        possible_frames = frames[len(frames) // 2:]
+        found = False
+        for frame_num in possible_frames:
+            best_box = longest_trace.get_box(frame_num)
+            best_pt = longest_trace.get_pt(frame_num)
+            best_labels = longest_trace.best_labels
+            best_scores = longest_trace.best_scores
+            best_frame = frame_num
+            if best_box is not None:
+                found = True
+                break
+
+        # Choose the first frame if no box is found
+        if not found:
+            best_frame = frames[0]
+            best_box = longest_trace.get_box(best_frame)
+            best_pt = longest_trace.get_pt(best_frame)
+            best_labels = longest_trace.best_labels
+            best_scores = longest_trace.best_scores
+
         if rescale:
-            pt, box = self.rescale(pt, box)
-        if frame_num in self.score.keys():
-            score = self.score[frame_num]
-        else:
-            score = 0.
-        return pt, self.best_label, box, score
+            best_box = self.rescale_box(best_box)
+            best_pt = self.rescale_pt(best_pt)
+        return best_frame, best_pt, best_labels, best_box, best_scores
+
+    def get(self, frame_num:int = -1, rescale=True) -> (int, np.array, str, np.array, float):
+        if frame_num == -1:
+            frame_num = list(self.pt.keys())[-3]
+
+        pt = None
+        for trace in self.traces:
+            if frame_num in trace.pt.keys():
+                pt = trace.get_pt(frame_num)
+                break
+
+        box = None
+        for trace in self.traces:
+            if frame_num in trace.box.keys():
+                box = trace.get_box(frame_num)
+                break
+
+        if rescale:
+            if box is not None:
+                box = self.rescale_box(box)
+            if pt is not None:
+                pt = self.rescale_pt(pt)
+
+        # Get the longest trace
+        longest_trace = self.traces[0]
+        for trace in self.traces:
+            if trace.num_frames > longest_trace.num_frames:
+                longest_trace = trace
+        best_labels = longest_trace.best_labels
+        best_scores = longest_trace.best_scores
+
+        return pt, best_labels[0], box, best_scores[0]
 
     def predict(self) -> np.array:
-        return self.pt[self.last_updated_frame]
+        predictions = []
+        for trace in self.traces:
+            pt = trace.get_pt(-1)
+            if pt is not None:
+                predictions.append(pt)
+        return np.array(predictions)
 
-    def update(self, label: str, pt: np.array, emb: np.array, frame_num: int, box:np.array = None, score:float = None) -> None:
+    def init(self, trace_id:int, label: str, pt: np.array, emb: np.array, frame_num: int, box:np.array = None, score:float = None) -> None:
         if self.is_closed():
-            debug(f"Tracker {self.id} has a gap from {self.last_updated_frame} to {frame_num} or more than max_frames {self.max_frames}")
+            info(f"{self.id} is closed")
             return
 
-        # If updating the same last_updated_frame, replace the point
-        if frame_num == self.last_updated_frame:
-            info(f"Updating tracker {self.id} at frame {frame_num} with point {pt}")
-            self.pt[frame_num] = pt
-            self.label[frame_num] = label
-            self.box[frame_num] = box
-            self.score[frame_num] = score
-            # If there is a valid embedding, update it
-            if len(emb) > 0:
-                self.emb = emb
-            return
+        self.traces[trace_id].update_pt_box(label, pt, frame_num, box, score)
 
-        # If adding in a new last_updated_frame, add the point
-        self.pt[frame_num] = pt
-        self.label[frame_num] = label
-        self.box[frame_num] = box
-        self.score[frame_num] = score
-        if len(emb) > 0:
-            self.emb = emb
-        self.last_updated_frame = frame_num
-
-        # Reduce the impact of the early detections by only considering the last 10 frames
-        scores = np.array(list(self.score.values()))
-        labels = list(self.label.values())
-        if len(scores) > 10:
-            scores = scores[-10:]
-            labels = labels[-10:]
-
-        # Calculate the weighted score for each label
-        label_scores = defaultdict(float)
-        for score, label in zip(scores, labels):
-            label_scores[label] += float(score)
-
-        # Update the best_label with a weighted probability
-        label = max(label_scores, key=label_scores.get)
-        num_best_label = labels.count(label)
-        self.best_label = label
-        self.best_score = label_scores[label] / num_best_label
-
-        pts_pretty = [f"{pt[0]:.2f},{pt[1]:.2f},{label},{score}" for pt, label, score in zip(self.pt.values(), self.label.values(), self.score.values())]
-        total_frames = len(self.pt)
-        info(f"Updating tracker {self.id} total_frames {total_frames} updated start {self.start_frame} to {self.last_updated_frame} {pts_pretty} with label {self.best_label}, score {self.best_score}")
+        # Require an embedding to update_pt_box the tracker
+        assert len(emb) > 0, "Embedding is required to initialize a tracker"
+        self.emb = emb
+        self.close_check(frame_num)